Enhances blocks integration with dynamic labels
Improves the PawaPay payment gateway's block integration by using dynamic labels for form fields and messages.

This change allows for easier customization and translation of the payment form within the WooCommerce Blocks interface.
It also adds a fallback to the default translation if a specific label is not defined.

Additionally, disables SSL verification for API requests.